backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.1
deepgram-sdk>=3.1.0
openai>=1.12.0
python-multipart>=0.0.9
alembic>=1.13.1
